requirements.txt
kafka-python
pandas
scikit-learn
torch
numpy
pyspark==3.5.3
cassandra-driver
graphframes
